MVP - full code refactored
change the code architecture - moved mappers calls from controllers into services

+++ src/main/java/com/training/studienplaner/submission/Submission.java
@@ -49,5 +49,5 @@ public class Submission {
     private Status status = Status.NOT_SUBMITTED;
 
     @Column
-    private Short grade;
+    private Double grade;
 }
